move access transformers implementation to access-inl.h
This is necessary to localize changes for Payload-specific functions in the
single place.
